Fix #127557: Polyline gesture creates extra point
This commit adds a bool to signal the user's intent so that dependent
code can determine the number of points the gesture should have.

In comparison with the original PR, this commit also updates the
behavior in 4.2 to display the click target similarly to 4.3 to avoid
confusing behavior now that clicking inside the circle is not a valid
action with only two points.

Original PR: !127827