Fix #128914: GPv3: Crash applying modifier to all frames
The issue was that the `merge_layers` function assumed that the
`src_grease_pencil` always returns a drawing when calling
`get_eval_drawing` which can return `nullptr`.
Another issue was that it tried to use `new_grease_pencil` to
fetch a `src_drawing`.

The fix makes sure to check that the `src_drawing` exists
and uses `src_grease_pencil` to fetch the drawing.