DRW: Add split visibility and command generation function
This adds the possibility to the user code to
group visibility and command generation compute
dispatch together to reduce the overhead of
PSO and pipeline switch.

This PR also make passes that have been submitted
(i.e.: that generated commands) as read-only.

This allows to automatically remove visibility
and command generation when submitting the same
pass with the same view multiple times or the
same view in multiple passes.

TODO:
- Fix invalid usage throughout the code
- Add tests cases.