Fix #126939: Smooth / Enhance Details brush crash when using tablet
Prior to this commit, the brush action was determined on a step by step
basis by peeking at the `StrokeCache` `bstrength` property and
determining which brush to use based on if the value was negative or
not.

The sign of this value, however, is not static across the entirety of a
brush stroke, as it is calculated from three separate fields, one of
which can vary over the course of a stroke when using a tablet, the
`pen_flip` property.

To fix this issue, this commit adds a new boolean to store the initial
direction of the stroke at the beginning of the stroke.