Make asset activation and active highlight work
This was tricky, basically we need to handle interaction with asset weak
references well. Idea is that weak references keep using the original
essentials path, so they remain working when stored in files and used on
other machines. Only when resolving the references to file paths, we
also resolve them to the override location, if it exists.

In some cases we also need to restore the original essential asset path
from the path to the overrides.