ffmpeg: Support output of 10- and 12-bit videos with relevant codecs
Part of task #118493:

When ffmpeg video with suitable codec is selected, the regular
image output UI shows color depth choice. That is:
- 10 bit option for H.264, H.265, AV1
- 12 bit option for H.265, AV1
VP9 could do 10 bit in theory too, but ffmpeg that is built in
Blender does not have that compiled in.

When "lossless" is picked, then similar to how for regular 8-bit video
it switches from YUV 4:2:0 to 4:4:4, this also switches to 4:4:4 10-
or 12-bit variant.