Fix #126273: Retiming movies with mismatched FPS is broken
There were too many issues to list them individually. Mostly they stem
from not correctly translating timeline frames to frame index domains.

Where it should or should not be used is not super straight forward.
This fix is not very nice either and no attempt to make this code nicer
was made.