Sequencer: add remaining time estimate to renderinfo string
As requested here: https://blender.community/c/rightclickselect/RMqK/

From a user perspective, this change increases the visibility of a very
useful number by placing it where it feels most intuitive.

From a developer perspective, this change puts information derived from
the current RenderJob into the infostring, when previously it only
contained information about the current Render (i.e. the current
frame). This is rather messy; as we now need
- to track the start time inside RenderJob
- a flag for whether or not the job is a sequence render job
- a carve-out inside the infostring callback to calculate the estimate
  in the sole case of the sequence flag being true.

Alternatives explored:
Calculating the estimate inside do_render_sequencer() and modifying the
RenderStats infostr was my first thought (similar to how engines add
time estimates for single frames), but since that is within the
animation loop and does not have access to the RenderJob, and a
timestamp for the sequence render start is not (easily) available
there. (there is WM_jobs_starttime, but that depends on there being a
wmWindowManager).

Upsides:
This change could be extended to add a time estimate to the renderinfo
string for non-sequencer animation renders too.

If, in the future, the sequencer were made a top-level item, i feel it
would simplify this code as well.