UI: Enable Region Overlap in Node Editor
Make sidebar see-through, like the 3D Viewport.

I believe this was not enabled back in the days due to the issue
Region Overlap had of not being able to click through the empty area.

This has been solved a while ago, so now it is more valuable to be
able to see what's behind the sidebar.