Fix: UI: Assets catalog sorting in asset shelf mixes catalogs
Sorting assets by catalog was introduced for the asset shelf in
471378c666. Intention was to keep related/similar assets closer
together. But the sorting was based on the catalog name only, so
the catalog order didn't make much sense, and closer related assets
would still end up being separated by less related ones.

Instead the full catalog path should be compared, so that the hierarchy
is reflected better, and for example sibling catalogs follow each other.
This way related assets are actually placed in close proximity, as
initially intended.

(This patch also makes the asset browser sort assets by catalogs by
default, which isn't exposed in the search options UI yet. Makes it
easier to test the effects of this PR)