Fix #129420: precision issue in light tree distant light angle
In volume segment, the minimal angle formed by the bounding cone axis
and the vector pointing from the cluster centroid and any point on the
ray is computed via `dot(bcone.axis, point_to_centroid)`, see Fig.8. in
paper.
For distant light this angle is 0, but due to numerical issues this is
not always true. Therefore explicitly assign `-bcone.axis` to
`point_to_centroid` in this case.