Fix #129353: GPv3: Crash when using Catmul-Rom curve type
Introduced by 5fff95f5193ed22a5c06462bc073095eafdd53b4.
The issue was that the `edit_points_vflag` buffer is of size `total_points_num`
which is based on the number of control points not evaluated positions.
The fix is to use `points_by_curve` instead of `points_by_curve_eval`
to get the right indices.