Fix #128954: Effect strip can not be easily blended with its input
Since #38598, inputs of render strips are not rendered, unless they are
positioned above the effect strip. This is to correctly render
transition effects with transparent inputs.

Current setup allows to create some effects, such as text outline.
However blurring part of the image is much more complicated. You have
to position input for blur effect above effect strip, and choose alpha
under blending. Then effect strip can be cropped. This is quite
unintuitive.

With this commit, input of effect strip, which is positioned below the
effect strip is not rendered only when the effect has 2 inputs.

Downside of this change is, that correctly blurring transparent image
requires muting of blur effect input.