GPv3: Add back Alt constraint to lines
This is invoked by holding Alt when starting the stroke.

Unlike existing `paint_stroke_line_constrain` this only allows a single
straight line without odd behaviour when releasing ALT.

- This uses a small distance instead of first sample to determine direction.
This required the need to store the samples until this metric was exceeded.
This stores the samples in a deferred_sample vector until the direction
is established.
- Move PaintStroke to paint_intern.hh to allow event to be stored and recalled.