Fix #129425: Using Voxel Remesh in sculpt mode can crash
It is possible for the relevant `SculptSession` data, specifically the
PVBH, to be in an invalid state when running the voxel remesh
operation. When inside sculpt mode, this can cause a crash by
dereferencing a nullptr when attempting to push the geometry state.

To fix this, this commit adds a call to
`BKE_sculpt_update_object_for_edit` to ensure that this data is
populated.