Nodes: Add "Power" and "Sign" operations to Vector Math Node
This adds "Power" and "Sign" as per-element/channel operations to the Vector Math node.

From my experience, it's a fairly common occurence to have to split vectors, apply these operations to each component, and combine them again. So it would be useful if these could be done directly with the Vector Math node instead.