Fix broken palette updates when unified paint is enabled.
Some modes and tools want to use the Brush.color property instead of
UnifiedPaintSettings.color. This does not work because the brush color
update will always look at the global flag. Add an explicit version of
the color update that works with brush colors. Check the context mode
in the event handler to know whether the ups flag should be ignored.