Nodes: Change isotropic 3D Gabor noise random orientations
This patch changes the random orientations methods for the isotropic 3D
Gabor noise. The old method simply adds random angles to the base
orientation in polar coordinates, but this generates non-uniform
distributions and is biased in direction.

The new method performs uniform sphere picking on a unit sphere then
orients the sphere in the direction of the base orientation. However, it
also suffers from discontinuities and rolling since there is no
reference direction for the orientation.