Fix #129726: UV islands selection can be very slow
When introduced in bb89cc51eca3, island selection for box/circle/lasso
was providing a `UvNearestHit`  -- more importantly a `BMFace` to
`uv_select_linked_multi` for every face that was touched by selection.
This can be very slow if many faces/islands are in play.

Fortunately, `uv_select_linked_multi` does not need to have `BMFace`
provided, it can work on an existing selection and select from there we
we can make use of for a nice performance gain. However, **deselection**
would not work well for this, so we need to keep the old/slow method for
deselection.