Fix: Failing boolean diff test with recent Xcode
Newer Xcode (compared to the one used on buildbot) changed something
that affects precision, which makes it hard to guarantee indices in
specific tests.

It wasn't very practical from time investment perspective to try to
align precision between different Xcode versions, so instead allow
meshes to have different indices in the failing test that was failing.