Fix: Suppress UBSAN division by zero warning for PBVH raycast function
This commit suppresses the division by zero warning when the ray being
cast has zero for any directional component. This case is expected in
this algorithm, as it depends on the underlying implementation being
IEEE 754 floating point properties (i.e. that 1/0 => positive infinity)

See [1] for more background reading on the paper this algorithm was
originally based on

[1] An Efficient and Robust Ray–Box Intersection Algorithm
    https://people.csail.mit.edu/amy/papers/box-jgt.pdf