Deps: add Protobuf + its dependency Abseil
Add Google Protobuf (https://protobuf.dev/) and its dependency Abseil.

Protobuf is a likely candidate to be used for the Online Assets
Repository index file.

In this PR, Protobuf is built as a shared library, with its dependency
Abseil statically linked in.

Protobuf has a second dependency, `utf8_range`, which comes bundled with
the Protobuf sources. This is automatically built using the same choice
for static/dynamic linking as Protobuf itself, so I haven't managed to
get that to link statically into the Protobuf library as well.

Source modules that want to use Protobuf can use `${Protobuf_INCLUDE_DIRS}`
in their include paths, and `PRIVATE ${Protobuf_LITE_LIBRARIES}` in their
libraries. I want to stick to the "lite" build of Protobuf for Blender
itself.

The Protobuf code generator (`protoc`) needs the regular Protobuf library,
so in the precomiled libs we need both 'lite' and 'regular' flavours.