USD: Add :object_name / :data_name metadata on import
When importing an USD, the Blender object names do not necessarily match Prim names. This makes it difficult in the [USD Import Hooks](https://docs.blender.org/api/current/bpy.types.USDHook.html) to link a Blender object bak to its source Prim for additional processing. A typical use cases for games is to generate UVs, create and apply a material on the fly when importing a collision shape that does not have a visual representation (hence no materials) based on some Prim attributes, but that the artist needs to differenciate in Blender's viewport.
This MR injects the custom `userProperties:blender:object_name` and `userProperties:blender:data_name` string attributes (the same attributes that are set when exporting an USD) during import, making them available in the import hooks.

Co-authored-by: Odréanne Breton <odreanne.breton@ubisoft.com>
Co-authored-by: Sttevan Carnali Joga <sttevan.carnali-joga@ubisoft.com>
Co-authored-by: Charles Flèche <charles.fleche@ubisoft.com>