USD: Allow to merge transform and shape on export
Adds the option `merge_transform_and_shape` to the USD export operator and panels so the transform and shapes are merged into a single USD Prim. Without the option, a top-level mesh would be exported as a top-level `Xform` that has a `Mesh` child. This matches the Blender data model, where a transformable object contains a geometric shape (like a mesh). This structure is also very valid in USD, where we don't want to directly instanciate geometric primitives. However, "since number of prims on a stage is one of the primary factors that governs how USD scale" (see https://openusd.org/release/glossary.html#usdglossary-instancing ), to reduce the number of prims in a stage, geometric primitives *are transformable* themselves.

Co-authored-by: Odréanne Breton <odreanne.breton@ubisoft.com>
Co-authored-by: Sttevan Carnali Joga <sttevan.carnali-joga@ubisoft.com>
Co-authored-by: Charles Flèche <charles.fleche@ubisoft.com>
As such, the new export option allows to export geometric primitives without the parent transform: